Reorganizacion de views
