+++ tecnicas/views/__init__.py
@@ -1,16 +1,20 @@
 from .autentication import autentication
 from .login_catador import catadorLogin
 from .main_panel import mainPanel
-from .management_catadores import managementCatadores
-from .sessions_panel import sesionsPanel
-from .seleccion_tecnica import selecionTecnica
-from .configuration_panel_basic import configurationPanelBasic
-from .configuration_panel_tags import configurationPanelTags
-from .configuration_panel_codes import configurationPanelCodes
-from .configuration_panel_words import configurationPanelWords
-from .api_tag import newTag
-from .tester_create import testerCreate
-from .tester_search import testerSearch
-from .api_words import words
-from .create_session import createSession
-from .session_details import sessionDetails
+
+from .sessions_management.sessions_panel import sesionsPanel
+from .sessions_management.session_details import sessionDetails
+
+from .sessions_config.seleccion_tecnica import selecionTecnica
+from .sessions_config.configuration_panel_basic import configurationPanelBasic
+from .sessions_config.configuration_panel_tags import configurationPanelTags
+from .sessions_config.configuration_panel_codes import configurationPanelCodes
+from .sessions_config.configuration_panel_words import configurationPanelWords
+from .sessions_config.create_session import createSession
+
+from .tester_management.tester_menu import testerMenu
+from .tester_management.tester_create import testerCreate
+from .tester_management.tester_search import testerSearch
+
+from .apis.api_tag import newTag
+from .apis.api_words import words
