+++ tecnicas/views/tester_management/tester_menu.py
@@ -1,4 +1,4 @@
 from django.shortcuts import render
 
-def managementCatadores(req):
+def testerMenu(req):
     return render(req, "tecnicas/manage_tester/catadores-panel.html")
